pytorch-transformer/requirements.txt
NumPy >= 1.11.1
Pytorch >= 0.3.0
nltk
tensorboardX